Split GIF frames intelligently
Lsix has the nice feature that if you type "lsix foo.gif" it will show
you all the frames within the GIF as thumbnails.

However, if a person types "lsix" at the command line, without
arguments, they don't expect whatever random GIF files to be expanded
in the same way.

This commit fixes it so that only the first frame of GIFs are shown if
lsix is called with no arguments.