Adjust #tiles to window size, handle redirection, add shadows
Three changes.

1. Now uses dtterm escape sequences to get the window size (if allowed)
and calculates the proper number of tiles to show. Default width is
set at the top of the file.

There are now variables for the default tile size as well. At the
moment, I've left it at 120x120 which is ImageMagick's default.

Since we're specifying the tile size, I now had to also specify the
space around each tile. I guessed that margin around each picture
should about 1% of the width of the screen and that seems right. (0.5%
on left and right). I made the margin above and below half of that
because, well, it looked right.

2. Escape sequences for terminal autodetection are always sent to
stderr instead of stdout. This means, I can now do 'lsix > foo.sixel'
to save a montage. (ImageMagick will let you also convert it to other
formats. E.g., 'convert foo.sixel foo.png')

3. If you have a white background terminal you'll notice that each
thumbnail now has a nice little drop shadow behind it. It even
properly handles transparency in PNGs and SVG files. It's a subtle
touch, but I think it's worth the extra half-a-second to calculate.