Elliding long filenames
I got tired of Montage always overwriting its own titles when the
filenames were long. This branch is an experiment of removing matching
prefixes and suffixes. It is not Unicode aware, but it does kind of
work.

Given files named superlongfilenameomgthisiscrazy01.png to
superlongfilenameomgthisiscrazy99.png, it'd show the full title for
only the first, and then the unique part for the following files:

	superlongfilenameomgthisiscrazy01.png
	...02
	...03
	...04

Remaining bugs: Does not handle spaces. Have not checked for dealing
with shell escape codes. Does not do unicode. Lines which are still
too long after elliding ought to be properly split with a newline.