Remove bug (long lines) and add feature (wrapping)
Montage is silly in that it does not wrap text even though it knows
the tile size and the proportional font size. Lsix handles it by
simply assuming if it's over 15 characters long, it's gotta be
trimmed. (Heuristically works, but needs more testing).

Could be improved by looking for letters nearby the break point and
wrapping on spaces, dashes and the like.