Use inline image resize (foo.jpg[>120x>120]) for speed
This may be premature optimization, but I think it makes a noticeable
difference for large numbers of files.

ImageMagick docs say that it's much faster to resize the file
immediately when loading it and much less resource intensive. This
makes sense as it won't have to malloc as much.

It wasn't hard to add. All you need is a suffix after the filename
like "foo.jpg[120x120]" to specify the width and heigth. Since I only
want to shrink images that are larger than 120x120, not magnify small
ones, I use the greater than sign: "foo.jpg[>120x>120]"

Of course, I also didn't hardcode 120, but instead used $tilewidth and
$tileheight, which are set at the beginning of the file.

Note that this [120x120] suffix seems to be mutually exclusive with
the [0] used for selecting the first frame from a GIF. Since GIFs not
explicitly mentioned on the command line only show the first frame,
those GIFs will be loaded full size.