fix slow start bug with thousands of images
When thousands of images were specified on the command line, lsix
would take a long time to startup. The problem was that we were
running the processlabel command to wordwrap all the filenames. Now,
one row at a time is processed, as originally intended.

This does show that, in the future, I should look at optimizing
process label somewhat.