use `readarray` since Apple is silly
In the days of yore, the Bash `mapfile` command was once called `readarray`.

Apple, for no good reason, ships with an ancient version of Bash which
doesn't understand `mapfile`. I'd prefer to not use the `env` kludge to
run a user installed bash, so I'm reverting `mapfile` to the archaic
`readarray` form so that MacOS users can enjoy lsix without any fuss.