Allow GIF files to have a colon in the filename.
ImageMagick does not allowing a colon in a filename if we are
selecting a frame by suffixing [0] to a .gif. E.g.,

	  foo:bar.gif

is treated as:

	  foo:bar.gif[0]

and ImageMagick fails to find the file "bar.gif".

The solution is to use a colon at the beginning of the filename to
tell Imagemagick to treat colons literally. So, we'd treat the file as

	:foo:bar.gif[0]

Ugly, yes. But, the little gods of ImageMagick must be appeased.