Add more exceptions for yaft
Yaft, being a VT102 terminal, cannot respond to VT220 queries, for
example if it sixel support, how many colors it can show, and what its
foreground and background colors are. Fortunately, all of those are
hardcoded in yaft, so we can just check $TERM == yaft* in a few
places. Its a bit ugly having exceptions, but I think one is okay.