Allow broken terminals to set LSIX_FORCE_SIXEL_SUPPORT
Ideally, terminals should respond properly to the Send Device
Attributes request to let us know that they support sixels. However,
we do not live in an ideal world. While I'd like this kludge to be
used as little as possible -- the terminal should have a bug report
filed and get patched or I should add an exception in lsix -- I'm
adding this variable to save some people the time of having to hack
the lsix script to remove the hassixel test.