GIFs with colons fail: ImageMagick >=6.9.10 requires "file://"
If a GIF filename has a colon in it, say "11:11.gif", we have to add a
prefix to tell ImageMagick to not interpret the colon. This happens
because ImageMagick acts funny when we suffix "[0]" at the end of the
filename, as we do with GIFs, to pick just the first frame.

The previous solution (commit 0535c2be1c2dd2b156cfffad65d50ad6bdf07f3a)
was to prefix ":" before any GIF files. That worked well in older
versions of ImageMagick, but no longer works in 6.9.10. The new
solution is to prefix "file://". To prevent the prefix from being
displayed in the montage, we add it in after the -label has been generated.

The full filename for "11:11.gif" becomes "file://11:11.gif[0]".

Note that for simplicity we add the prefix to *all* files, not just
GIFs, but it should (hopefully) not have any harmful effects.