Presume sixel scrolling is enabled
After purchasing a genuine DEC VT340 terminal, hackerb9 has
learned that the DEC VT340 graphics programmer's manual was
mistaken and the sense of DECSDM should be inverted. Nearly
every terminal emulator which built upon that reference over
the last thirty years is incorrect. Some are correcting the
error now, some never were wrong, but we are at a time when
it is not safe to presume what setting DECSDM means.

For now, lsix will simply presume that sixel scrolling is
enabled and hope that no program has been run before it that
disables it.

See https://github.com/hackerb9/lsix/issues/41 for details.

Thanks go to @j4james for pointing out the mistake and for
testing multiple terminal emulators to show that none of
them currently default to having sixel scrolling disabled.