Workaround Microsoft Windows' builtin "convert" command
Please see pull request #42 from @StaticPH for complete details.

Note that while something like 'alias convert="magick convert"' is
theoretcially possible, it actually fails on some machines, such as
the current stable version of Debian, since "magick" only comes with
the most recent versions of ImageMagick.

Instead, we'll do a workaround presuming that (a) montage is a unique
name that only ImageMagick uses, and (b) the convert binary is in the
same directory.

We check $COMSPEC under the belief that that environment variables
gets set if (and only if) the operating system is Microsoft Windows.

I dislike the extra messiness caused by these five lines of code, but
I guess it cannot be helped if `lsix` is to work under Microsoft
Windows.