shopt -s expand_aliases is now necessary when non-interactive
Apparently bash changed that behaviour at some point. It's handy to be
able to use aliases to patch up missing features and quirks in a way
that doesn't muck with the rest of the script, so I wonder why bash
has it disabled by default now. Perhaps people had been abusing it in
bigger scripts, leading to confusion and difficulty debugging.

That's not an issue for lsix, fortunately.