Simpler workaround
Simpler isn't always better, but for `lsix`, simpler is a primary
goals.

This code adds the flaw that `magick` (a binary that comes with newer
versions of ImageMagick) is required on Microsoft Windows machines but
not on any other OS. That works for me since I don't run Microsoft
Windows, but I hope it doesn't inconvenience too many other folk.