alias montage="magick montage" for IM7
Montage is not deprecated, yet, but it looks like it is going to be.

Also, switched to using an alias (instead of a variable) for convert
as it looks cleaner. Downside is that if they don't have a sixel
terminal, the error message will tell users to type `convert` even for
IM7. For now, that's fine.

Perhaps of note about the return value of `type` and `command`:

       This command       |  Is equivalent to these commands
    ----------------------|----------------------------------
       type foo           |  command -v foo
       type foo bar       |  type foo  &&  type bar
       command -v foo bar |  command -v foo  ||  command -v bar

Therefore, `command -v magick montage` does an implicit "OR", which is
what we want when checking if either of IM6 or IM7 is installed.